Modified firehose schema by adding Context info
Before, the context was embedded in location in the report, for instance
by adding a function name field in the Location info.

To provide much more scalability across checkers and ease future
modifications, two new "elements/classes" are there Context and
Declaration.

Context represents what happened around the result reported : this can
be in a declaration (more generic than function : name and kind) and/or
the checker provide a trace that itself includes a list of States :
Declaration/Location/Notes/Annotations.

Test cases has been rewritten for this modification.